Refactor legal pages to reflect that they are 'informational' legal pages
Informational as in, only for information purposes. There is no agree/disagree option.
