+++ client/src/layout/legal/informational/DisclaimerPage.js
@@ -3,13 +3,13 @@ import {
   Header
 } from 'semantic-ui-react';
 
-import PageLayout from '../components/PageLayout';
+import PageLayout from '../../components/PageLayout';
 import LegalTextContainer from './LegalTextContainer';
 
-const TermsOfServicePage = () => (
+const DisclaimerPage = () => (
   <PageLayout>
     <LegalTextContainer>
-      <Header as='h1'>Terms of Service</Header>
+      <Header as='h1'>Disclaimer</Header>
       <Header as='h2'>Section</Header>
       <p>
         Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
@@ -26,4 +26,4 @@ const TermsOfServicePage = () => (
   </PageLayout>
 )
 
-export default TermsOfServicePage;
+export default DisclaimerPage;
