+++ client/src/layout/legal/informational/LegalTextContainer.js
